Added offset outline to action submission buttons
Added the outline-offset class to the timeline action submission buttons to improve
their visual clarity. And added specifier for focus pseudo-class to offset-outline
class to increase its specificity. Without this change in specificity, the class's
outline-color property would be overridden by the one given to button:focus in modal.css