Added outlines to toggle switches
This branch combines changes from button-outlines and
rename-hidden-css-class. The CSS for toggle switches has been modified
so that there is an outline when one is in focus. Currently this
outline forms around the label and the switch as two connected boxes,
rather than being enclosed by a single rectangle. This is less
noticeable at smaller font sizes.